+++ src/onegov.gazette/onegov/gazette/tests/test_layout.py
@@ -50,7 +50,7 @@ def test_layout_links():
     assert layout.manage_users_link == '/UserCollection/'
     assert layout.manage_user_groups_link == '/UserGroupCollection/'
     assert layout.manage_notices_link == '/GazetteNoticeCollection/'
-    assert layout.manage_published_notices_link == '/GazetteNoticeCollection/'
+    assert layout.manage_accepted_notices_link == '/GazetteNoticeCollection/'
     assert layout.manage_statistics_link == \
         '/GazetteNoticeCollection/statistics/'
     assert layout.dashboard_link == '/dashboard/'
@@ -65,13 +65,13 @@ def test_layout_menu():
     request._is_personal = True
     assert layout.menu == [
         ('My Drafted and Submitted Official Notices', '/dashboard/', False),
-        ('My Published Official Notices', '/GazetteNoticeCollection/', False)
+        ('My Accepted Official Notices', '/GazetteNoticeCollection/', False)
     ]
 
     request._is_private = True
     assert layout.menu == [
         ('Submitted Official Notices', '/GazetteNoticeCollection/', False),
-        ('Published Official Notices', '/GazetteNoticeCollection/', False),
+        ('Accepted Official Notices', '/GazetteNoticeCollection/', False),
         ('Statistics', '/GazetteNoticeCollection/statistics/', False)
     ]